Move dpack to serialtbx.detector.cspad
