+++ xfel/cxi/cspad_ana/cspad_tbx.py
@@ -24,6 +24,7 @@ import six
 from six.moves import zip
 
 from serialtbx import util
+import serialtbx.detector.cspad
 
 __version__ = "$Revision$"
 
@@ -450,82 +451,9 @@ def CsPadElement(data3d, qn, config):
     quadrant = numpy.rot90(quadrant, 4 - qn)
   return quadrant
 
-def dpack(active_areas=None,
-          address=None,
-          beam_center_x=None,
-          beam_center_y=None,
-          ccd_image_saturation=None,
-          data=None,
-          distance=None,
-          pixel_size=pixel_size,
-          saturated_value=None,
-          timestamp=None,
-          wavelength=None,
-          xtal_target=None,
-          min_trusted_value=None):
-  """XXX Check completeness.  Should fill in sensible defaults."""
-
-  # Must have data.
-  if data is None:
-    return None
-
-  # Create a time stamp of the current time if none was supplied.
-  if timestamp is None:
-    timestamp = evt_timestamp()
-
-  # For unknown historical reasons, the dictionary must contain both
-  # CCD_IMAGE_SATURATION and SATURATED_VALUE items.
-  if ccd_image_saturation is None:
-    if saturated_value is None:
-      ccd_image_saturation = cspad_saturated_value
-    else:
-      ccd_image_saturation = saturated_value
-  if saturated_value is None:
-    saturated_value = ccd_image_saturation
-
-  # Use a minimum value if provided for the pixel range
-  if min_trusted_value is None:
-    min_trusted_value = cspad_min_trusted_value
-
-  # By default, the beam center is the center of the image.  The slow
-  # (vertical) and fast (horizontal) axes correspond to x and y,
-  # respectively.
-  if beam_center_x is None:
-    beam_center_x = pixel_size * data.focus()[1] / 2
-  if beam_center_y is None:
-    beam_center_y = pixel_size * data.focus()[0] / 2
-
-  # By default, the entire detector image is an active area.  There is
-  # no sensible default for distance nor wavelength.  XXX But setting
-  # wavelength to zero may be disastrous?
-  if active_areas is None:
-    # XXX Verify order with non-square detector
-    active_areas = flex.int((0, 0, data.focus()[0], data.focus()[1]))
-  if distance is None:
-    distance = 0
-  if wavelength is None:
-    wavelength = 0
-
-  # The size must match the image dimensions.  The length along the
-  # slow (vertical) axis is SIZE1, the length along the fast
-  # (horizontal) axis is SIZE2.
-  return {'ACTIVE_AREAS': active_areas,
-          'BEAM_CENTER_X': beam_center_x,
-          'BEAM_CENTER_Y': beam_center_y,
-          'CCD_IMAGE_SATURATION': ccd_image_saturation,
-          'DATA': data,
-          'DETECTOR_ADDRESS': address,
-          'DISTANCE': distance,
-          'PIXEL_SIZE': pixel_size,
-          'SATURATED_VALUE': saturated_value,
-          'MIN_TRUSTED_VALUE': min_trusted_value,
-          'SIZE1': data.focus()[0],
-          'SIZE2': data.focus()[1],
-          'TIMESTAMP': timestamp,
-          'SEQUENCE_NUMBER': 0, # XXX Deprecated
-          'WAVELENGTH': wavelength,
-          'xtal_target': xtal_target}
-
+def dpack(*kwargs):
+  """ thin wrapper """
+  return serialtbx.detector.cspad.dpack(*kwargs)
 
 def hdf5pack(hdf5_file,
              active_areas=None,
